Add additional documentation for Notifiers (#530)
* Add additional documentation for Notifiers

When I first set up the notifiers, I felt like the documentation was pretty light on how to do this. 

I'd like to submit a few more PRs to outline a couple basic examples on this page (or a separate page specific for notifications) in addition to elaborating more on how this feature works. It's super flexible, and I love that!

Nevertheless, I think it would be more accessible to outline a few examples on what this ends up entailing.

* User general/latest shoutrrr URL without hardcoded version